[ SFT QA ] moving FILE_CHUNK message generator from itest specific place to common place to make it usable for unit tests too.
